Added detailed explaination of the the project purpose, some background for M&M use and technical details for the R implementation.
In the future the code chunks should be broken down into several R scripts to prevent the document from failing to render if a single component breaks.

I was unsure how to address which output file should be used for the yaml so it was removed (the markdown is just used for inline viewing of graph and table outputs rather than kniting an output document).

Finally, on my new laptop this diretory was located under a onedrive file structure so i changed the location to a local file path.